つけ忘れてたownerplugin追加とお金をdoubleに

+++ src/main/java/dev/felnull/Data/StorageData.java
@@ -10,11 +10,11 @@ public class StorageData {
     public String groupName;//グループ名(個人プレイヤーもグループ扱い)
     public GroupData groupData;//ストレージの所属グループ
     public final Set<String> requireBankPermission;//ストレージ直下の金庫の要求パーミッション(BukkitPermではない)
-    public int bankMoney;//ストレージ付属金庫の値
+    public double bankMoney;//ストレージ付属金庫の値
     @NotNull
     public Map<String,InventoryData> storageInventory;//ストレージに含まれているインベントリデータ　キーのStringはページ名
 
-    public StorageData(@NotNull Set<String> requireBankPermission, Map<String,InventoryData> storageInventory, int bankMoney) {
+    public StorageData(@NotNull Set<String> requireBankPermission, Map<String,InventoryData> storageInventory, double bankMoney) {
         this.requireBankPermission = requireBankPermission;
         if(storageInventory == null){
             storageInventory = new HashMap<>();
